wasm: preserve 32-bit validator layout
Why:
- uint32 frame stamps consumed memory outside padding on 32-bit targets

What:
- use compact uint16 epochs with bounded rollover clearing
- enforce unchanged validator and control-frame sizes on 32-bit and 64-bit targets

Proof:
- go test ./src/core/compiler/wasm
- GOARCH=386 focused tests, 100 repetitions
- focused race tests, Linux/arm64 cross-build
- differential fuzzing: 896,907 executions
- BenchmarkDecodeValidate: 80.9 us/op, 48,999 B/op, 359 allocs/op

Docs:
- unchanged, internal validator footprint fix with no workflow impact

+++ src/core/compiler/wasm/validate.go
@@ -878,7 +878,7 @@ type ctrlFrame struct {
 	ifSeenElse   bool
 	// branchTableEpoch uses existing tail padding to mark whether this exact
 	// label frame has already been checked by the current br_table.
-	branchTableEpoch uint32
+	branchTableEpoch uint16
 }
 
 type funcValidator struct {
@@ -904,7 +904,7 @@ type funcValidator struct {
 	constOnly         bool
 	// branchTableEpoch fits in the padding before constGlobalLimit. Each
 	// funcValidator owns its control frames, including under parallel validation.
-	branchTableEpoch uint32
+	branchTableEpoch uint16
 	constGlobalLimit int // globals below this absolute index are visible to a const expression
 	// rd is reused across bodies validated by this funcValidator so the byte
 	// cursor is not heap-allocated per function/const-expression.
